bin/ico2gltf/requirements.txt
kaitaistruct~=0.8
Pillow~=7.2.0
gltflib~=1.0
